ML Endpoint/requirement.txt
flask==3.1.0
flask-cors==5.0.0
tensorflow==2.9.1
numpy==1.24.4
pillow==11.0.0
gunicorn==23.0.0